requirements.txt
requests~=2.28.2
beautifulsoup4~=4.12.2
googlesearch-python
mysql-connector-python~=8.0.32
pandas~=2.0.0
sqlalchemy~=2.0.11
pymongo~=4.3.3
neo4j~=5.7.0
dash~=2.9.2
plotly~=5.14.1
mysql~=0.0.3